requirements.txt
psycopg2-binary==2.9.9
django==5.1.2
pika==1.3.2

